requirements.txt
dj-database-url==0.5.0
gunicorn==19.7.1
whitenoise==3.3.1
psycopg2==2.7.2
django-heroku==0.3.1
mysql-connector-python==8.0.11
mysqlclient==1.3.10


cachecontrol >= 0.12.5
google-auth >= 1.5.0
google-cloud-firestore >= 0.29.0
google-cloud-storage >= 1.10.0
google-cloud-core >= 0.28.1
cachetools >= 2.1.0
firebase-admin >= 2.11.0
google-api-core >= 1.2.0
google-resumable-media >= 0.3.1
googleapis-common-protos >= 1.5.3
msgpack >= 0.5.6
protobuf >= 3.5.2
pyasn1 >= 0.4.3 
pyasn1-modules >= 0.2.1 
rsa >=3.4.2

requests>=2.9.1

django-cors-headers >= 2.2.0